+++ base-plus-starter/src/main/java/org/gourd/hu/base/config/FilterConfig.java
@@ -4,14 +4,12 @@ import org.gourd.hu.base.request.filter.RequestDetailFilter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
 /**
  * 过滤器配置
  *
  * @author gourd.hu
  */
-@Configuration
 public class FilterConfig {
 
     @Autowired
