+++ cloud-plus-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  org.gourd.hu.base.config.BaseAutoConfig
+  org.gourd.hu.cloud.config.CloudAutoConfig
