--- core-plus-starter/src/main/java/org/gourd/hu/CoreConfiguration.java
@@ -1,14 +1,0 @@
-package org.gourd.hu;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.retry.annotation.EnableRetry;
-
-/**
- * @author gourd.hu
- * @date 2018-11-20
- */
-@Configuration
-@EnableRetry
-public class CoreConfiguration {
-
-}
