+++ document-plus-starter/src/main/java/org/gourd/hu/doc/config/FileAutoConfig.java
@@ -1,6 +1,8 @@
-package org.gourd.hu;
+package org.gourd.hu.doc.config;
 
 import com.github.tobato.fastdfs.FdfsClientConfig;
+import org.gourd.hu.doc.fastdfs.FastDfsClient;
+import org.gourd.hu.doc.openoffice.utils.FileConvertUtil;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableMBeanExport;
 import org.springframework.context.annotation.Import;
@@ -11,8 +13,8 @@ import org.springframework.jmx.support.RegistrationPolicy;
  * @date 2018-11-20
  */
 @Configuration
-@Import(FdfsClientConfig.class)
+@Import({FdfsClientConfig.class, FileConvertUtil.class, FastDfsClient.class})
 @EnableMBeanExport(registration = RegistrationPolicy.IGNORE_EXISTING)
-public class FileConfiguration {
+public class FileAutoConfig {
 
 }
