+++ document-plus-starter/src/main/java/org/gourd/hu/doc/openoffice/utils/FileConvertUtil.java
@@ -3,7 +3,6 @@ package org.gourd.hu.doc.openoffice.utils;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.CollectionUtils;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
 import java.io.BufferedReader;
@@ -16,7 +15,6 @@ import java.util.List;
 /**
  * @author gourd
  */
-@Component
 @Slf4j
 public class FileConvertUtil {
