+++ es-plus-starter/src/main/java/org/gourd/hu/es/repository/UserRepository.java
@@ -2,10 +2,12 @@ package org.gourd.hu.es.repository;
 
 import org.gourd.hu.es.model.entity.UserEs;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
+import org.springframework.stereotype.Component;
 
 /**
  * @author gourd
  */
+@Component
 public interface UserRepository extends ElasticsearchRepository<UserEs,Long> {
