+++ mongo-plus-starter/src/main/java/org/gourd/hu/mongo/config/MongoAutoConfig.java
@@ -1,13 +1,16 @@
 package org.gourd.hu.mongo.config;
 
+import org.gourd.hu.mongo.service.impl.UserMogServiceImpl;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 /**
  *
  * @author gourd.hu
  */
 @Configuration
-public class MongoConfig {
+@Import({UserMogServiceImpl.class})
+public class MongoAutoConfig {
 
     /**
      * 副本集模式支持
