+++ quartz-plus-starter/src/main/java/org/gourd/hu/quartz/job/DemoJob.java
@@ -3,7 +3,6 @@ package org.gourd.hu.quartz.job;
 import lombok.extern.slf4j.Slf4j;
 import org.quartz.*;
 import org.springframework.scheduling.quartz.QuartzJobBean;
-import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
 
@@ -11,7 +10,6 @@ import java.time.LocalDateTime;
  * job触发时间
  * @author gourd
  */
-@Component
 @Slf4j
 @PersistJobDataAfterExecution
 @DisallowConcurrentExecution
